+++ src/BasicDisplays/index.ts
@@ -2,3 +2,4 @@ export * from './BooleanDisplay';
 export * from './DatetimeDisplay';
 export * from './RawDisplay';
 export * from './ModelDisplay';
+export * from './EnumDisplay';