+++ src/BasicInputs/stories/DatetimeInput.stories.tsx
@@ -5,13 +5,13 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { DatetimeInput } from '../DatetimeInput';
 
 const meta = {
-  title: 'Commons/DateTime',
+  title: 'Commons/BasicInputs/DateTime',
   tags: ['autodocs'],
   render: () => {
     const [date, setDate] = useState<string | undefined | null>(
       '2024-07-10T23:12:34',
     );
-    return <DatetimeInput value={date} onChange={setDate} hourCycle={24} />;
+    return <DatetimeInput value={date} onChange={setDate} hourCycle={12} />;
   },
 } satisfies Meta<typeof DatetimeInput>;
 export default meta;